requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
djangorestframework==3.13.1
idna==3.3
python-dotenv==0.20.0
pytz==2022.1
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
